+++ includes/admin/templates/layout/title.php
@@ -6,6 +6,7 @@ use WP_STATISTICS\Admin_Template;
 use WP_Statistics\Utils\Request;
 use WP_Statistics\Components\View;
 
+View::load('components/objects/usage-tracker-notice');
 ?>
 <div class="wps-wrap__top <?php echo isset($real_time_button) ? 'wps-wrap__top--has__realtime' : ''; ?>">
     <?php if (isset($backUrl, $backTitle)): ?>