+++ includes/admin/templates/settings/advanced.php
@@ -200,21 +200,21 @@ add_thickbox();
                 <input id="geoip_license_key" type="text" size="30" name="wps_geoip_license_key" value="<?php echo esc_attr(WP_STATISTICS\Option::get('geoip_license_key')); ?>">
                 <p class="description">
                     <?php
-                        /* translators: %s: Link to maxmind */
-                        echo sprintf(
-                            wp_kses(
-                                __('Enter your <strong>MaxMind license key</strong> to enable the <strong>premium MaxMind GeoIP database</strong>, which provides more precise location data. The plugin uses the free database by default. %s', 'wp-statistics'),
-                               [
+                    /* translators: %s: Link to maxmind */
+                    echo sprintf(
+                        wp_kses(
+                            __('Enter your <strong>MaxMind license key</strong> to enable the <strong>premium MaxMind GeoIP database</strong>, which provides more precise location data. The plugin uses the free database by default. %s', 'wp-statistics'),
+                            [
                                 'strong' => [],
                                 'a'      => [
                                     'href'   => [],
                                     'class'  => [],
                                     'target' => [],
                                 ]
-                               ]
-                            ),
-                            '<a href="https://www.maxmind.com/en/geoip-databases" class="wps-text-decoration-underline" target="_blank">' . esc_html__('Get MaxMind Premium.', 'wp-statistics') . '</a>'
-                        );
+                            ]
+                        ),
+                        '<a href="https://www.maxmind.com/en/geoip-databases" class="wps-text-decoration-underline" target="_blank">' . esc_html__('Get MaxMind Premium.', 'wp-statistics') . '</a>'
+                    );
                     ?>
                 </p>
             </td>
@@ -228,22 +228,22 @@ add_thickbox();
                 <input id="geoip_dbip_license_key_option" type="text" size="30" name="wps_geoip_dbip_license_key_option" value="<?php echo esc_attr(WP_STATISTICS\Option::get('geoip_dbip_license_key_option', '')); ?>">
                 <p class="description">
                     <?php
-                        /* translators: %s: Link to dbip */
-                        echo sprintf(
-                            wp_kses(
-                                __('Enter your DB-IP license key to enable the premium DB-IP database, replacing the free version with a more detailed dataset.<br /> The premium DB-IP database is <strong>1.1GB</strong> in size. Make sure your server has enough storage space before enabling it, as the plugin downloads and stores this database. %s', 'wp-statistics'),
-                                [
-                                    'br'     => [],
-                                    'strong' => [],
-                                    'a'      => [
-                                        'href'   => [],
-                                        'class'  => [],
-                                        'target' => [],
-                                    ]
+                    /* translators: %s: Link to dbip */
+                    echo sprintf(
+                        wp_kses(
+                            __('Enter your DB-IP license key to enable the premium DB-IP database, replacing the free version with a more detailed dataset.<br /> The premium DB-IP database is <strong>1.1GB</strong> in size. Make sure your server has enough storage space before enabling it, as the plugin downloads and stores this database. %s', 'wp-statistics'),
+                            [
+                                'br'     => [],
+                                'strong' => [],
+                                'a'      => [
+                                    'href'   => [],
+                                    'class'  => [],
+                                    'target' => [],
                                 ]
-                            ),
-                            '<a href="https://db-ip.com/db/?refid=vrn" class="wps-text-decoration-underline" target="_blank">' . esc_html__('Get DB-IP Premium.', 'wp-statistics') . '</a>'
-                        );
+                            ]
+                        ),
+                        '<a href="https://db-ip.com/db/?refid=vrn" class="wps-text-decoration-underline" target="_blank">' . esc_html__('Get DB-IP Premium.', 'wp-statistics') . '</a>'
+                    );
                     ?>
                 </p>
             </td>
@@ -315,7 +315,7 @@ add_thickbox();
                 function handle_geoip_fields() {
                     var method = jQuery("#geoip_location_detection_method").val();
                     var isMaxmind = (method === "maxmind");
-                    var isDBIP    = (method === "dbip");
+                    var isDBIP = (method === "dbip");
                     var isUserLicense = (jQuery("#geoip_license_type").val() === "user-license");
 
                     jQuery("#geoip_license_type_option, #enable_geoip_option, #schedule_geoip_option, #geoip_auto_pop_option")
@@ -444,6 +444,26 @@ add_thickbox();
                 <p class="description"><?php esc_html_e('Revert all user-specific and global configurations to the WP Statistics default settings, preserving your existing data.', 'wp-statistics'); ?></p>
                 <p class="description"><span class="wps-note"><?php esc_html_e('Caution', 'wp-statistics'); ?>:</span> <?php esc_html_e('This change is irreversible.', 'wp-statistics'); ?></p>
                 <p class="description"><?php _e('<b>For multisite users</b>: Every site within the network will return to the default settings.', 'wp-statistics'); // phpcs:ignore WordPress.Security.EscapeOutput.UnsafePrintingFunction	?></p>
+            </td>
+        </tr>
+        </tbody>
+    </table>
+</div>
+
+<div class="postbox">
+    <table class="form-table">
+        <tbody>
+        <tr valign="top">
+            <th scope="row" colspan="2"><h3><?php esc_html_e('Usage Tracking', 'wp-statistics'); ?></h3></th>
+        </tr>
+        <tr valign="top">
+            <th scope="row">
+                <label for="wps_usage_data_tracking"><?php esc_html_e('Enable Usage Tracking', 'wp-statistics'); ?></label>
+            </th>
+            <td>
+                <input id="wps_usage_data_tracking" type="checkbox" name="wps_usage_data_tracking" <?php echo WP_STATISTICS\Option::get('usage_data_tracking') == true ? "checked='checked'" : ''; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>'>
+                <label for="wps_usage_data_tracking"><?php esc_html_e('Enable', 'wp-statistics'); ?></label>
+                <p class="description"><?php esc_html_e('Allow WP Statistics to collect anonymous usage data to improve features, performance, and user experience. No personal or sensitive data will be collected. Learn more.', 'wp-statistics'); ?></p>
             </td>
         </tr>
         </tbody>